add utils/arg_group.py for the common args
+ move the commonly shared groups of arguments parsing functions from utils/plot.py into a new sub-module utils/arg_group.py

+ move the parallel computing related arguments into utils/arg_group.add_parallel_argument()

+ adjust usages across the whole mintpy package
