+++ mintpy/plot_transection.py
@@ -11,7 +11,8 @@ import sys
 import argparse
 import numpy as np
 from matplotlib import pyplot as plt, ticker
-from mintpy.utils import readfile, utils as ut, plot as pp
+
+from mintpy.utils import arg_group, readfile, utils as ut, plot as pp
 from mintpy import view
 
 
@@ -68,8 +69,8 @@ def create_parser():
     lines.add_argument('--ms', '--markersize', dest='marker_size', type=float, default=2.0,
                        help='Point marker size. Default: 2.0')
 
-    parser = pp.add_figure_argument(parser)
-    parser = pp.add_save_argument(parser)
+    parser = arg_group.add_figure_argument(parser)
+    parser = arg_group.add_save_argument(parser)
     return parser
 
 
@@ -153,8 +154,8 @@ def get_view_cmd(iargs):
     parser = argparse.ArgumentParser(description='view.py parser')
     parser.add_argument('--noverbose', dest='print_msg', action='store_false',
                         help='Disable the verbose message printing.')
-    parser = pp.add_figure_argument(parser)
-    parser = pp.add_save_argument(parser)
+    parser = arg_group.add_figure_argument(parser)
+    parser = arg_group.add_save_argument(parser)
 
     # get args that are applicable to view.py
     unique_args = parser.parse_known_args(iargs)[1]
